cleanup warning in stage3.cpp

+++ src/stage3.cpp
@@ -34,7 +34,7 @@ void shellDisplayResults(const slvr::Solver& solver) {
             if (input == "help" || input == "h") {
                 std::cout << msgs::stage3CommandsMessage << std::endl;
             } else if (input == "validate" || input == "va") {
-                if (numSolutions == solutionsVector.size() && thread_numSolutions == thread_solutionsVector.size()) {
+                if (numSolutions == (int)solutionsVector.size() && thread_numSolutions == (int)thread_solutionsVector.size()) {
                     std::cout << "Looks good!" << std::endl;
                 } else {
                     std::cerr << "Uh oh, looks like the solution sizing is off" << std::endl;
